refactor codebase, use ens-avatar lib
