+++ src/config.ts
@@ -12,8 +12,7 @@ const CANVAS_FALLBACK_FONT_PATH = path.join(FONT_FOLDER, 'DejaVuSans-Bold.ttf');
 const INAMEWRAPPER = process.env.INAMEWRAPPER || '0x1aa28a1e';
 
 const INFURA_API_KEY = process.env.INFURA_API_KEY;
-const IPFS_GATEWAY = 'https://ipfs.io/ipfs/';
-const IPNS_GATEWAY = 'https://ipfs.io/ipns/';
+const IPFS_GATEWAY = 'https://ipfs.io/';
 
 const ADDRESS_ETH_REGISTRAR = process.env.ADDRESS_ETH_REGISTRAR || '0x57f1887a8BF19b14fC0dF6Fd9B2acc9Af147eA85';
 const ADDRESS_NAME_WRAPPER = process.env.ADDRESS_NAME_WRAPPER || '0x4D83cea620E3864F912046b73bB3a6c04Da75990';
@@ -30,6 +29,5 @@ export {
   INAMEWRAPPER,
   INFURA_API_KEY,
   IPFS_GATEWAY,
-  IPNS_GATEWAY,
   SERVER_URL,
 };
