+++ src/service/domain.ts
@@ -7,7 +7,7 @@ import {
 } from './subgraph';
 import { Metadata } from './metadata';
 import { getAvatarImage } from './avatar';
-import { Version } from './base';
+import { Version } from '../base';
 
 const eth =
   '0x93cdeb708b7545dc668eb9280176169d1c33cfd8ed6f04690a0bcc88a93fc4ae';
