+++ src/service/rasterize.ts
