+++ src/utils/base64encode.ts
@@ -1,0 +1,9 @@
+const btoa = require('btoa');
+
+export default function base64EncodeUnicode(str: string): string {
+  return btoa(
+    encodeURIComponent(str).replace(/%([0-9A-F]{2})/g, function (_, p1) {
+      return String.fromCharCode(parseInt(p1, 16));
+    })
+  );
+}
