+++ test/avatar.test.ts
@@ -1,97 +1,112 @@
 import avaTest, { ExecutionContext, TestInterface } from 'ava';
 import { TestContext } from './interface';
-import { IPFS_GATEWAY, IPNS_GATEWAY } from '../src/config';
-import { AvatarMetadata } from '../src/avatar';
+import { IPFS_GATEWAY } from '../src/config';
+import { utils } from '@ensdomains/ens-avatar'
 
 const test = avaTest as TestInterface<TestContext>;
 
 test('should return data URIs without any changes', async (t: ExecutionContext<TestContext>) => {
   const uri = 'data:text/plain;base64,dGVzdGluZw==';
-  t.is(uri, AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(uri, resolvedURI);
 });
 
 test('should return http URIs without any changes', async (t: ExecutionContext<TestContext>) => {
   const uri = 'https://app.ens.domains';
-  t.is(uri, AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(uri, resolvedURI);
 });
 
 test('should replace ipfs://ipfs/ with IPFS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = 'ipfs://ipfs/QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP';
+  const { uri: resolvedURI } = utils.resolveURI(uri);
   t.is(
-    IPFS_GATEWAY + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
-    AvatarMetadata.parseURI(uri)
+    IPFS_GATEWAY + 'ipfs/' + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
+    resolvedURI
   );
 });
 
 test('should replace ipfs:// with IPFS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = 'ipfs://QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP';
+  const { uri: resolvedURI } = utils.resolveURI(uri);
   t.is(
-    IPFS_GATEWAY + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
-    AvatarMetadata.parseURI(uri)
+    IPFS_GATEWAY + 'ipfs/' + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
+    resolvedURI
   );
 });
 
 test('should replace /ipfs/ with IPFS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = '/ipfs/QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP';
+  const { uri: resolvedURI } = utils.resolveURI(uri);
   t.is(
-    IPFS_GATEWAY + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
-    AvatarMetadata.parseURI(uri)
+    IPFS_GATEWAY + 'ipfs/' + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
+    resolvedURI
   );
 });
 
 test('should replace ipfs/ with IPFS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = 'ipfs/QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP';
+  const { uri: resolvedURI } = utils.resolveURI(uri);
   t.is(
-    IPFS_GATEWAY + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
-    AvatarMetadata.parseURI(uri)
+    IPFS_GATEWAY + 'ipfs/' + 'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP',
+    resolvedURI
   );
 });
 
 test('should recognize ipfs hash with subpath', async (t: ExecutionContext<TestContext>) => {
   const uri =
     'ipfs/QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP/avatar/name.jpg';
+  const { uri: resolvedURI } = utils.resolveURI(uri);
   t.is(
-    IPFS_GATEWAY +
+    IPFS_GATEWAY + 'ipfs/' + 
       'QmUbTVz1L4uEvAPg5QcSu8Pow1YdwshDJ8VbyYjWaJv4JP/avatar/name.jpg',
-    AvatarMetadata.parseURI(uri)
+    resolvedURI
   );
 });
 
 test('should prefix CIDs with IPFS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const cidv0 = 'QmbWqxBEKC3P8tqsKc98xmWNzrzDtRLMiMPL8wBuTGsMnR';
-  t.is(IPFS_GATEWAY + cidv0, AvatarMetadata.parseURI(cidv0));
+  const { uri: resolvedURIv0 } = utils.resolveURI(cidv0);
+  t.is(IPFS_GATEWAY + 'ipfs/' + cidv0, resolvedURIv0);
   const cidv1 = 'bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi';
-  t.is(IPFS_GATEWAY + cidv1, AvatarMetadata.parseURI(cidv1));
+  const { uri: resolvedURIv1 } = utils.resolveURI(cidv1);
+  t.is(IPFS_GATEWAY + 'ipfs/' + cidv1, resolvedURIv1);
 });
 
 test('should replace ipfs://ipns/ with IPNS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = 'ipfs://ipns/testing';
-  t.is(IPNS_GATEWAY + 'testing', AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(IPFS_GATEWAY + 'ipns/' + 'testing', resolvedURI);
 });
 
 test('should replace ipns://ipns/ with IPNS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = 'ipns://ipns/testing';
-  t.is(IPNS_GATEWAY + 'testing', AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(IPFS_GATEWAY + 'ipns/' + 'testing', resolvedURI);
 });
 
 test('should replace ipns:// with IPNS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   const uri = 'ipns://testing';
-  t.is(IPNS_GATEWAY + 'testing', AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(IPFS_GATEWAY + 'ipns/' + 'testing', resolvedURI);
 });
 
 test('should replace /ipns/ with IPNS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   // Should only replace the first occurence of /ipns/
   const uri = '/ipns/testing/ipns/other';
-  t.is(IPNS_GATEWAY + 'testing/ipns/other', AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(IPFS_GATEWAY + 'ipns/' + 'testing/ipns/other', resolvedURI);
 });
 
 test('should replace ipns/ with IPNS gateway prefix', async (t: ExecutionContext<TestContext>) => {
   // Should only replace the first occurence of ipns/
   const uri = 'ipns/testing/ipns/other';
-  t.is(IPNS_GATEWAY + 'testing/ipns/other', AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(IPFS_GATEWAY + 'ipns/' + 'testing/ipns/other', resolvedURI);
 });
 
 test('should return any URI that does not match any of the previous conditions unchanged', async (t: ExecutionContext<TestContext>) => {
   const uri = 'testing';
-  t.is(uri, AvatarMetadata.parseURI(uri));
+  const { uri: resolvedURI } = utils.resolveURI(uri);
+  t.is(uri, resolvedURI);
 });
